+++ pepselect-child/inc/setup.php
@@ -23,6 +23,7 @@ require_once get_stylesheet_directory() . '/inc/contact-page.php';
 require_once get_stylesheet_directory() . '/inc/military-page.php';
 require_once get_stylesheet_directory() . '/inc/account.php';
 require_once get_stylesheet_directory() . '/inc/checkout.php';
+require_once get_stylesheet_directory() . '/inc/emails.php';
 
 /**
  * Determine whether the required Hello Elementor parent theme is available.